Break up Makefile in sbatch-less and with sbatch. Include dependency test.
The sbatch parameters within the Makefile were not necessary for users without
sbatch. Therefore a separate Makefile-sbatch was created that calls the
Makefile. It comes from my other repository gnu-make-sbatch-tools. This commit
also creates a dependecy checker script to make sure all dependencies are met
to be able to run the pipeline.